Small capsule and capsule wall done
Small capsule and capsule wall done.

- Add | Wall parts notify the "wall holder" that it has been hit and
then the "wall holder" handles the collision. A mechanic also was added
to make sure that only one part hit was counted per frame, this might be
the cause of a bug where nothing sometimes happens to the the wall when
hit.
- Add | Fixed so that both the player and bullets can pass through the
holes in the wall.